% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_plot.R
\name{allocation_plot}
\alias{allocation_plot}
\title{Plot results of the \code{allocation()} function}
\usage{
allocation_plot(allocation, bb_area)
}
\arguments{
\item{allocation}{The output of the \code{\link[=allocation]{allocation}}
function.}

\item{bb_area}{A \code{\link[sf:st_as_sf]{sf}} boundary box object with the area
of interest.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2}} plot showing the potential locations
for new facilities.
}
\description{
\code{allocation_plot()} plot the results of the \code{\link[=allocation]{allocation}}
function, showing the potential locations for new facilities and the
coverage attained.
}
\examples{
\dontrun{
  traveltime_data <- traveltime(
    facilities = naples_fountains,
    bb_area = naples_shape,
    dowscaling_model_type = "lm",
    mode = "walk",
    res_output = 100
  )

  allocation_data <-
    naples_population |>
    allocation(
      traveltime = traveltime_data,
      bb_area = naples_shape,
      facilities = naples_fountains,
      weights = NULL,
      objectiveminutes = 15,
      objectiveshare = 0.99,
      heur = "max",
      dowscaling_model_type = "lm",
      mode = "walk",
      res_output = 100
    )

  allocation_data |> allocation_plot(naples_shape)
}
}
\seealso{
Other plot functions: 
\code{\link{allocation_plot_discrete}()}
}
\concept{plot functions}
\keyword{reporting}
