% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodata.R
\name{demo_data_load}
\alias{demo_data_load}
\title{Demo data for locationallocation}
\usage{
demo_data_load()
}
\description{
This function loads in the global environment four demo data objects to test the functions of the locationallocation package:
\itemize{
\item boundary: a sf polygon geometry object representing the administrative boundary of the city of Naples, Italy
\item fountains: a sf point geometry object representing the water fountains in the city of Naples, Italy
\item pop: a raster Raster object representing the population density in the city of Naples, Italy
\item hotdays: a raster Raster object representing the number of hot days in the city of Naples, Italy
}
}
\keyword{vignette}
