% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationallocation-package.R
\docType{package}
\name{locationallocation-package}
\alias{locationallocation}
\alias{locationallocation-package}
\title{locationallocation: Maximal Coverage Location-Allocation Using Geospatial Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions to perform Maximal Coverage Location-Allocation analyses using geospatial data. It is designed to help decision-makers optimally allocate resources and facilities in various contexts, such as public health, emergency response, and urban planning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://giacfalk.github.io/locationallocation/}
  \item \url{https://github.com/giacfalk/locationallocation/}
  \item Report bugs at \url{https://github.com/giacfalk/locationallocation/issues/}
}

}
\author{
\strong{Maintainer}: Giacomo Falchetta \email{giacomo.falchetta@cmcc.it} (\href{https://orcid.org/0000-0003-2607-2195}{ORCID}) [conceptor, copyright holder]

Other contributors:
\itemize{
  \item Daniel Vartanian \email{danielvartan@proton.me} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [contributor]
}

}
\keyword{internal}
