% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_raster_to_polygon.R
\name{mask_raster_to_polygon}
\alias{mask_raster_to_polygon}
\title{Mask a \code{stars} object to a \code{sf} object}
\usage{
mask_raster_to_polygon(ras, mask, inverse = FALSE, updatevalue = NA)
}
\arguments{
\item{ras}{A \code{\link[stars:st_as_stars]{stars}} object.}

\item{mask}{A \code{\link[stars:st_as_stars]{stars}} or \code{\link[sf:st_as_sf]{sf}}
object.}

\item{inverse}{A \code{\link[base:logical]{logical}} flag. If \code{TRUE}, the mask
is inverted (default: \code{FALSE}).}

\item{updatevalue}{The value to update the \code{\link[stars:st_as_stars]{stars}}
object with (default: \code{NA}).}
}
\value{
A masked \code{\link[stars:st_as_stars]{stars}} object.
}
\description{
This function rapidly masks a \code{\link[stars:st_as_stars]{stars}} object to
a \code{\link[sf:st_as_sf]{sf}} object.
}
\examples{
naples_population |> mask_raster_to_polygon(naples_shape)
}
\concept{utility functions}
\keyword{masking}
