% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naples_fountains}
\alias{naples_fountains}
\title{Water fountains in the city of Naples, Italy}
\format{
A \code{\link[sf:st_as_sf]{sf}} \code{POINT} geometry object with 251
feature and 24 fields:
}
\usage{
naples_fountains
}
\description{
A \code{\link[sf:st_as_sf]{sf}} polygon geometry object representing the
locations of water fountains in the city of Naples, Italy.
}
\seealso{
Other datasets: 
\code{\link{naples_hot_days}},
\code{\link{naples_population}},
\code{\link{naples_shape}}
}
\concept{datasets}
\keyword{datasets}
