% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naples_hot_days}
\alias{naples_hot_days}
\title{Hot days in the city of Naples, Italy}
\format{
A \code{\link[stars:st_as_stars]{stars}} object with 1 layer.
}
\source{
\href{https://www.urban-climate.eu/model}{UrbClim}.
}
\usage{
naples_hot_days
}
\description{
A \code{\link[stars:st_as_stars]{stars}} object representing the number of hot
days in the city of Naples, Italy.

This 100-meter resolution heat hazard map shows the number of days with
\href{https://en.wikipedia.org/wiki/Wet-bulb_globe_temperature}{Wet-Bulb Globe Temperature}
above 25 °C during 2008–2017, based on simulations from the
\href{https://www.urban-climate.eu/model}{UrbClim} model.
}
\seealso{
Other datasets: 
\code{\link{naples_fountains}},
\code{\link{naples_population}},
\code{\link{naples_shape}}
}
\concept{datasets}
\keyword{datasets}
