% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naples_population}
\alias{naples_population}
\title{Population density in the city of Naples, Italy}
\format{
A \code{\link[stars:st_as_stars]{stars}} object with 1 layer.
}
\source{
Global Human Settlement Layer
(\href{https://human-settlement.emergency.copernicus.eu}{GHSL}).
}
\usage{
naples_population
}
\description{
A \code{\link[stars:st_as_stars]{stars}} object representing the population
density in the city of Naples, Italy.

The dataset is based on the Global Human
Settlement Layer
(\href{https://human-settlement.emergency.copernicus.eu}{GHSL}) population grid
(\href{https://human-settlement.emergency.copernicus.eu/download.php?ds=pop}{GHS-POP}).
}
\seealso{
Other datasets: 
\code{\link{naples_fountains}},
\code{\link{naples_hot_days}},
\code{\link{naples_shape}}
}
\concept{datasets}
\keyword{datasets}
