% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naples_shape}
\alias{naples_shape}
\title{Administrative boundary of the city of Naples, Italy}
\format{
A \code{\link[sf:st_as_sf]{sf}} \code{MULTIPOLYGON} geometry object with 1
feature and 8 fields:

\describe{
\item{GISCO_ID}{
A \code{\link[base:character]{character}} field with the Geographic Information
System of the Commission
(\href{https://ec.europa.eu/eurostat/web/gisco}{GISCO}) ID of the
administrative unit.
}
\item{CNTR_CODE}{
A \code{\link[base:character]{character}} field with the country code of Italy.
}
\item{LAU_ID}{
A \code{\link[base:character]{character}} field with the Local Administrative
Unit (LAU) ID of the administrative unit.
}
\item{LAU_NAME}{
A \code{\link[base:character]{character}} field with the name of the
administrative unit.
}
\item{POP_2021}{
A \code{\link[base:double]{double}} field with the population of the
administrative unit as of 2021.
}
\item{POP_DENS_2021}{
A \code{\link[base:double]{double}} field with the population density (people per
square kilometer) of the administrative unit as of 2021.
}
\item{AREA_KM2}{
A \code{\link[base:double]{double}} field with the area (in square
kilometers) of the administrative unit.
}
\item{YEAR}{
A \code{\link[base:integer]{integer}} field with the year of the population
data (2021).
}
}
}
\usage{
naples_shape
}
\description{
A \code{\link[sf:st_as_sf]{sf}} polygon geometry object representing the
administrative boundary of the city of Naples, Italy.
}
\seealso{
Other datasets: 
\code{\link{naples_fountains}},
\code{\link{naples_hot_days}},
\code{\link{naples_population}}
}
\concept{datasets}
\keyword{datasets}
