% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltime_plot.R
\name{traveltime_plot}
\alias{traveltime_plot}
\title{Plot results of the \code{traveltime()} function}
\usage{
traveltime_plot(
  traveltime,
  bb_area,
  facilities = NULL,
  contour_traveltime = 15,
  annotation_location = "br",
  annotation_scale = TRUE,
  annotation_north_arrow = TRUE
)
}
\arguments{
\item{traveltime}{A \code{\link[base:list]{list}} with the output of the
\code{\link[=traveltime]{traveltime()}} function.}

\item{bb_area}{A \code{\link[sf:st_as_sf]{sf}} boundary box object with the area
of interest.}

\item{facilities}{A \code{\link[sf:st_as_sf]{sf}} object with the existing
facilities.}

\item{contour_traveltime}{(optional) A number indicating the contour
thresholds for the travel time (default: \code{15}).}

\item{annotation_location}{(optional) A \code{\link[base:character]{character}} string
specifying the location of the annotation on the plot. See
\code{\link[ggspatial:annotation_scale]{annotation_scale}} for possible values
(default: \code{"br"}).}

\item{annotation_scale}{(optional) A \code{\link[base:logical]{logical}} flag
indicating whether to include a scale annotation on the plot (default:
\code{TRUE}).}

\item{annotation_north_arrow}{(optional) A \code{\link[base:logical]{logical}} flag
indicating whether to include a north arrow annotation on the plot
(default: \code{TRUE}).}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2}} plot showing the travel time from
the  facilities to the area of interest.
}
\description{
\code{traveltime_plot()} plot the results of the \code{\link[=traveltime]{traveltime()}}
function, showing the travel time from the facilities to the area of
interest.
}
\examples{
\dontrun{
  traveltime_data <-
    naples_fountains |>
    traveltime(
      bb_area = naples_shape,
      dowscaling_model_type = "lm",
      mode = "walk",
      res_output = 100
    )

  traveltime_data |>
    traveltime_plot(
      bb_area = naples_shape,
      facilities = naples_fountains
    )
}
}
\seealso{
Other travel time functions: 
\code{\link{friction}()},
\code{\link{traveltime}()},
\code{\link{traveltime_stats}()}
}
\concept{travel time functions}
\keyword{reporting}
