% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltime_stats.r
\name{traveltime_stats}
\alias{traveltime_stats}
\title{Summarize results of the \code{traveltime()} function}
\usage{
traveltime_stats(
  traveltime,
  demand,
  objectiveminutes = 15,
  breaks = c(5, 10, 15, 30),
  print = TRUE
)
}
\arguments{
\item{traveltime}{A \code{\link[base:list]{list}} with the output of the
\code{\link[=traveltime]{traveltime()}} function.}

\item{demand}{A \code{\link[stars:st_as_stars]{stars}} object with the
demand layer (e.g. population density).}

\item{objectiveminutes}{(optional) A number indicating the target travel time
in minutes used to compute the statistics (default: \code{15}).}

\item{breaks}{(optional) A \code{\link[base:numeric]{numeric}} object indicating the
breaks (in minutes) for the cumulative curve plot
(default: \code{c(5, 10, 15, 30)}).}

\item{print}{(optional) A \code{\link[base:logical]{logical}} flag indicating
whether to print the results to the console (default: \code{TRUE}).}
}
\value{
An \link[base:invisible]{invisible} \code{\link[base:list]{list}} with the
following elements:
\itemize{
\item \code{coverage}: A \code{\link[base:numeric]{numeric}} value indicating the share of
demand covered within the objective travel time.
\item \code{unmet_demand}: A \code{\link[base:numeric]{numeric}} value indicating the share
of demand that remains unmet.
\item \code{data}: A \code{\link[dplyr:reexports]{tibble}} object with the data used to
generate the cumulative curve plot.
\item \code{plot}: A \code{\link[ggplot2:ggplot]{ggplot}} object with the cumulative curve
plot.
}
}
\description{
\code{traveltime_stats()} generates a summary of the
\code{\link[=traveltime]{traveltime()}} function results, producing a statistic for
the percent of demand which is covered within a given objective travel time
along with a cumulative curve plot.
}
\examples{
\dontrun{
  traveltime_data <-
    naples_fountains |>
    traveltime(
      bb_area = naples_shape,
      dowscaling_model_type = "lm",
      mode = "walk",
      res_output = 100
    )

  traveltime_data |>
    traveltime_stats(
      demand = naples_population,
      objectiveminutes = 15,
      breaks = c(5, 10, 15, 30)
    )
}
}
\seealso{
Other travel time functions: 
\code{\link{friction}()},
\code{\link{traveltime}()},
\code{\link{traveltime_plot}()}
}
\concept{travel time functions}
\keyword{reporting}
