% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{allocation}
\alias{allocation}
\title{Location allocation (continuous problem)}
\usage{
allocation(
  demand_raster,
  traveltime_raster = NULL,
  bb_area,
  facilities = facilities,
  weights = NULL,
  objectiveminutes = 10,
  objectiveshare = 0.99,
  heur = "max",
  dowscaling_model_type,
  mode,
  res_output,
  approach = "norm"
)
}
\arguments{
\item{demand_raster}{A raster object with the demand for the service.}

\item{traveltime_raster}{The output of the traveltime function. If not provided, the function will run the traveltime function first.}

\item{bb_area}{A boundary box object with the area of interest.}

\item{facilities}{A sf object with the existing facilities.}

\item{weights}{A raster with the weights for the demand.}

\item{objectiveminutes}{The objective travel time in minutes.}

\item{objectiveshare}{The share of the demand to be covered.}

\item{heur}{The heuristic approach to be used. Options are "max" (default) and "kd".}

\item{dowscaling_model_type}{The type of model used for the spatial downscaling of the travel time layer.}

\item{mode}{The mode of transport.}

\item{res_output}{The spatial resolution of the friction raster (and of the analysis), in meters. If <1000, a spatial downscaling approach is used.}

\item{approach}{The approach to be used for the allocation. Options are "norm" (default) and "equity". If "norm", the allocation is based on the normalized demand raster multiplied by the normalized weights raster. If "absweights", the allocation is based on the normalized demand raster multiplied by the raw weights raster.}
}
\description{
This function is used to allocate facilities in a continuous location problem. It uses the accumulated cost algorithm to find the optimal location for the facilities based on the demand, travel time, and weights for the demand, and target travel time threshold and share of the demand to be covered.
}
\keyword{location-allocation}
