% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_plot.R
\name{allocation_plot}
\alias{allocation_plot}
\title{Plot results of the allocation and allocation_discrete functions}
\usage{
allocation_plot(output_allocation, bb_area)
}
\arguments{
\item{output_allocation}{The output of the allocation or allocation_discrete functions.}

\item{bb_area}{A boundary box object with the area of interest.}
}
\description{
This function is used to plot the results of the allocation and allocation_discrete functions. It shows the potential locations for new facilities and the coverage attained.
}
\keyword{reporting}
