% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_plot.R
\name{allocation_plot_discrete}
\alias{allocation_plot_discrete}
\title{Plot results of the allocation and allocation_discrete functions}
\usage{
allocation_plot_discrete(output_allocation, bb_area)
}
\arguments{
\item{bb_area}{A boundary box object with the area of interest.}

\item{allocation_plot_discrete}{The output of the allocation function.}
}
\description{
This function is used to plot the results of the allocation and allocation_discrete functions. It shows the potential locations for new facilities and the coverage attained.
}
\keyword{reporting}
