% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friction.R
\name{friction}
\alias{friction}
\title{Download and optionally downscale the friction surface layer to conduct location-allocation analysis.}
\usage{
friction(bb_area, mode, res_output, dowscaling_model_type)
}
\arguments{
\item{bb_area}{A spatial object representing the area of interest.}

\item{mode}{The mode of transport. Options are "fastest" and "walk".}

\item{res_output}{The spatial resolution of the friction raster (and of the analysis), in meters. If <1000, a spatial downscaling approach is used.}

\item{dowscaling_model_type}{The type of model used for the spatial downscaling of the friction layer. Options are "lm" (linear model) and "rf" (random forest).}
}
\description{
This function retrieves the friction surface layer from the Malaria Atlas Project database and optionally downscales it to the spatial resolution of the analysis using road network data from OpenStreetMap. The function returns the friction surface layer, the transition matrix, and the geocorrection matrix.
}
\keyword{cats}
