% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_raster_to_polygon.R
\name{mask_raster_to_polygon}
\alias{mask_raster_to_polygon}
\title{Quick masking of a Rast object to an sf object}
\usage{
mask_raster_to_polygon(
  ras = NULL,
  mask = NULL,
  inverse = FALSE,
  updatevalue = NA
)
}
\arguments{
\item{ras}{A Rast object.}

\item{mask}{A Rast or sf object.}

\item{inverse}{Logical. If TRUE, the mask is inverted.}

\item{updatevalue}{The value to update the Rast object with.}
}
\description{
This function rapidly masks a Rast object to an sf object.
}
\keyword{masking}
