% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naples_hot_days}
\alias{naples_hot_days}
\title{Hot days in the city of Naples, Italy}
\format{
A \code{\link[raster:raster]{RasterLayer}} object with 1 layer.
}
\usage{
naples_hot_days
}
\description{
A \code{\link[raster:raster]{RasterLayer}} object representing the number of hot
days in the city of Naples, Italy.
}
\seealso{
Other datasets: 
\code{\link{naples_fountains}},
\code{\link{naples_population}},
\code{\link{naples_shape}}
}
\concept{datasets}
\keyword{datasets}
