% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naples_population}
\alias{naples_population}
\title{Population density in the city of Naples, Italy}
\format{
A \code{\link[raster:raster]{RasterLayer}} object with 1 layer.
}
\usage{
naples_population
}
\description{
A \code{\link[raster:raster]{RasterLayer}} object representing the population
density in the city of Naples, Italy.
}
\seealso{
Other datasets: 
\code{\link{naples_fountains}},
\code{\link{naples_hot_days}},
\code{\link{naples_shape}}
}
\concept{datasets}
\keyword{datasets}
