% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltime.R
\name{traveltime}
\alias{traveltime}
\title{Generate a traveltime map}
\usage{
traveltime(facilities, bb_area, dowscaling_model_type, mode, res_output = 100)
}
\arguments{
\item{facilities}{A sf object with the existing facilities.}

\item{bb_area}{A boundary box object with the area of interest.}

\item{dowscaling_model_type}{The type of model used for the spatial downscaling of the travel time layer.}

\item{mode}{The mode of transport.}

\item{res_output}{The spatial resolution of the friction raster (and of the analysis), in meters. If <1000, a spatial downscaling approach is used.}
}
\description{
This function generates a traveltime map based on the input facilities, bounding box area, and travel mode.
}
\keyword{cats}
