% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltime.R
\name{traveltime}
\alias{traveltime}
\title{Generate a travel time map}
\usage{
traveltime(
  facilities,
  bb_area,
  mode = "walk",
  dowscaling_model_type = "lm",
  res_output = 100,
  cache = FALSE,
  file = NULL
)
}
\arguments{
\item{facilities}{A \code{\link[sf:st_as_sf]{sf}} object with the existing
facilities.}

\item{bb_area}{A \code{\link[sf:st_as_sf]{sf}} boundary box object with the area
of interest.}

\item{mode}{(optional) A \code{\link[base:character]{character}} string indicating
the mode of transport. Options are \code{"fastest"} and \code{"walk"} (default =
\code{"walk"}).
\itemize{
\item For \code{"fastest"}: The friction layer accounts for multiple modes of
transport, including walking, cycling, driving, and public transport, and are
based on the Malaria Atlas Project (2019) \emph{Global Travel Speed Friction
Surface}.
\item For \code{"walk"}: The friction layer accounts only for walking speeds and is
based on the Malaria Atlas Project (2015) \emph{Global Walking Only Friction
Surface}.
}}

\item{dowscaling_model_type}{(optional) A \code{\link[base:character]{character}}
string indicating the type of model used for the spatial downscaling of the
friction layer. Options are \code{"lm"} (linear model) and \code{"rf"} (random
forest) (default: \code{"lm"}).}

\item{res_output}{(optional) A positive
\link[checkmate:checkIntegerish]{integerish} number indicating the spatial
resolution of the friction layer (and of the analysis), in meters. If the
resolution is less than \code{1000}, a spatial downscaling approach is used
(default: \code{100}).}

\item{cache}{(optional) A \code{\link[base:logical]{logical}} flag
indicating whether to cache the downloaded friction data for future use
(default: \code{TRUE}).}

\item{file}{(optional) A \code{\link[base:character]{character}} string indicating
the path to a local friction surface layer file. If provided, the
function will use this file instead of downloading the friction data.}
}
\value{
An \link[base:invisible]{invisible} \code{\link[base:list]{list}} with the
following elements:
\itemize{
\item \code{travel_time}: A \code{\link[stars:st_as_stars]{stars}} object with the travel
time map.
\item \code{friction}: A \code{\link[base:list]{list}} with the outputs of the
\code{\link[=friction]{friction()}} function.
}
}
\description{
\code{traveltime()} generates a travel time map based on the input facilities,
bounding box area, and travel mode.

See the \code{\link[=friction]{friction()}} function for details on how the friction
layer is generated.
}
\examples{
\dontrun{
  library(dplyr)

  traveltime_data <-
    naples_fountains |>
    traveltime(
      bb_area = naples_shape,
      dowscaling_model_type = "lm",
      mode = "walk",
      res_output = 100
    )

  traveltime_data |> glimpse()

  traveltime_data |>
    traveltime_plot(
      bb_area = naples_shape,
      facilities = naples_fountains
  )
}
}
\seealso{
Other travel time functions: 
\code{\link{friction}()},
\code{\link{traveltime_plot}()},
\code{\link{traveltime_stats}()}
}
\concept{travel time functions}
\keyword{cats}
