% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltime_plot.R
\name{traveltime_plot}
\alias{traveltime_plot}
\title{Plot results of the \code{traveltime()} function}
\usage{
traveltime_plot(
  traveltime,
  bb_area,
  facilities = NULL,
  contour_traveltime = 15
)
}
\arguments{
\item{traveltime}{A \code{\link[base:list]{list}} with the output of the
\code{\link[=traveltime]{traveltime()}} function.}

\item{bb_area}{A \code{\link[sf:st_as_sf]{sf}} boundary box object with the area
of interest.}

\item{facilities}{A \code{\link[sf:st_as_sf]{sf}} object with the existing
facilities.}

\item{contour_traveltime}{(optional) A number indicating the contour
thresholds for the travel time (default: \code{15}).}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2}} plot showing the travel time from
the  facilities to the area of interest.
}
\description{
\code{traveltime_plot()} is used to plot the results of the
\code{\link[=traveltime]{traveltime()}} function. It shows the travel time from the
facilities to the area of interest.
}
\examples{
\dontrun{
  traveltime_data <-
    naples_fountains |>
    traveltime(
      bb_area = naples_shape,
      dowscaling_model_type = "lm",
      mode = "walk",
      res_output = 100
    )

  traveltime_data |>
    traveltime_plot(
      bb_area = naples_shape,
      facilities = naples_fountains
    )
}
}
\seealso{
Other travel time functions: 
\code{\link{friction}()},
\code{\link{traveltime}()},
\code{\link{traveltime_stats}()}
}
\concept{travel time functions}
\keyword{reporting}
