% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltime_stats.r
\name{traveltime_stats}
\alias{traveltime_stats}
\title{Summary of traveltime function results}
\usage{
traveltime_stats(
  traveltime,
  demand_raster,
  breaks = c(5, 10, 15, 30),
  objectiveminutes = 15
)
}
\arguments{
\item{traveltime}{The output object of the traveltime function.}

\item{demand_raster}{The demand raster layer.}

\item{breaks}{The breaks (minutes) for the cumulative curve plot.}

\item{objectiveminutes}{The objective travel time in minutes to calculate the statistic from.}
}
\description{
This function generates a summary of the traveltime function results, producing a cumulative curve plot and a statistic for the fraction of demand which is covered within a given objective travel time.
}
\keyword{reporting}
